Separate out raw WebSocket protocol handling
