+++ tests/load.py
@@ -8,7 +8,7 @@ given a sequence number. Any errors are reported and counted.
 
 import sys, os, select, random, time, optparse, logging
 sys.path.insert(0,os.path.join(os.path.dirname(__file__), ".."))
-from websockify.websocket import WebSocketServer, WebSocketRequestHandler
+from websockify.websocketserver import WebSocketServer, WebSocketRequestHandler
 
 class WebSocketLoadServer(WebSocketServer):
 
@@ -35,12 +35,10 @@ class WebSocketLoad(WebSocketRequestHandler):
         self.send_cnt = 0
         self.recv_cnt = 0
 
-        try:
-            self.responder(self.request)
-        except:
-            print "accumulated errors:", self.errors
-            self.errors = 0
-            raise
+        self.responder(self.request)
+
+        print "accumulated errors:", self.errors
+        self.errors = 0
 
     def responder(self, client):
         c_pend = 0
@@ -62,7 +60,7 @@ class WebSocketLoad(WebSocketRequestHandler):
                     print err
 
                 if closed:
-                    self.send_close()
+                    break
 
             now = time.time() * 1000
             if client in outs:
