+++ websockify/websocketproxy.py
@@ -17,7 +17,7 @@ except: from SocketServer import ForkingMixIn
 try:    from http.server import HTTPServer
 except: from BaseHTTPServer import HTTPServer
 import select
-from websockify import websocket
+from websockify import websocketserver
 from websockify import auth_plugins as auth
 try:
     from urllib.parse import parse_qs, urlparse
@@ -25,7 +25,9 @@ except:
     from cgi import parse_qs
     from urlparse import urlparse
 
-class ProxyRequestHandler(websocket.WebSocketRequestHandler):
+class ProxyRequestHandler(websocketserver.WebSocketRequestHandler):
+
+    buffer_size = 65536
 
     traffic_legend = """
 Traffic Legend:
@@ -86,9 +88,11 @@ Traffic Legend:
             msg += " (using SSL)"
         self.log_message(msg)
 
-        tsock = websocket.WebSocketServer.socket(self.server.target_host,
-                                                 self.server.target_port,
-                connect=True, use_ssl=self.server.ssl_target, unix_socket=self.server.unix_target)
+        tsock = websocketserver.WebSocketServer.socket(self.server.target_host,
+                                                       self.server.target_port,
+                                                       connect=True,
+                                                       use_ssl=self.server.ssl_target,
+                                                       unix_socket=self.server.unix_target)
 
         self.request.setsockopt(socket.SOL_TCP, socket.TCP_NODELAY, 1)
         if not self.server.wrap_cmd and not self.server.unix_target:
@@ -217,7 +221,7 @@ Traffic Legend:
                 cqueue.append(buf)
                 self.print_traffic("{")
 
-class WebSocketProxy(websocket.WebSocketServer):
+class WebSocketProxy(websocketserver.WebSocketServer):
     """
     Proxy traffic to and from a WebSockets client to a normal TCP
     socket server target.
@@ -270,7 +274,7 @@ class WebSocketProxy(websocket.WebSocketServer):
                 "REBIND_OLD_PORT": str(kwargs['listen_port']),
                 "REBIND_NEW_PORT": str(self.target_port)})
 
-        websocket.WebSocketServer.__init__(self, RequestHandlerClass, *args, **kwargs)
+        websocketserver.WebSocketServer.__init__(self, RequestHandlerClass, *args, **kwargs)
 
     def run_wrap_cmd(self):
         self.msg("Starting '%s'", " ".join(self.wrap_cmd))
@@ -465,7 +469,7 @@ def websockify_init():
         if len(args) > 2:
             parser.error("Too many arguments")
 
-    if not websocket.ssl and opts.ssl_target:
+    if not websocketserver.ssl and opts.ssl_target:
         parser.error("SSL target requested and Python SSL module not loaded.");
 
     if opts.ssl_only and not os.path.exists(opts.cert):
